+++ app/src/androidTestEreferrals/java/org/eyeseetea/malariacare/data/sync/exporter/PushUseCaseShould.java
@@ -4,22 +4,42 @@ import static junit.framework.Assert.fail;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.Mockito.when;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
 
+import org.eyeseetea.malariacare.AssetsFileReader;
+import org.eyeseetea.malariacare.BuildConfig;
+import org.eyeseetea.malariacare.data.database.datasources.CountryVersionLocalDataSource;
+import org.eyeseetea.malariacare.data.database.datasources.SettingsDataSource;
+import org.eyeseetea.malariacare.data.database.datasources.SurveyLocalDataSource;
 import org.eyeseetea.malariacare.data.database.datasources.UserAccountDataSource;
 import org.eyeseetea.malariacare.data.database.model.OrgUnitDB;
 import org.eyeseetea.malariacare.data.database.model.ProgramDB;
 import org.eyeseetea.malariacare.data.database.model.SurveyDB;
 import org.eyeseetea.malariacare.data.database.model.UserDB;
+import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
+import org.eyeseetea.malariacare.data.repositories.OrganisationUnitRepository;
+import org.eyeseetea.malariacare.data.repositories.ProgramRepository;
+import org.eyeseetea.malariacare.domain.boundary.executors.IAsyncExecutor;
+import org.eyeseetea.malariacare.domain.boundary.executors.IMainExecutor;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IAppInfoRepository;
 import org.eyeseetea.malariacare.domain.boundary.repositories.ICredentialsRepository;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IDeviceRepository;
+import org.eyeseetea.malariacare.domain.boundary.repositories.IOrganisationUnitRepository;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IProgramRepository;
+import org.eyeseetea.malariacare.domain.boundary.repositories.ISettingsRepository;
+import org.eyeseetea.malariacare.domain.boundary.repositories.ISurveyRepository;
+import org.eyeseetea.malariacare.domain.entity.AppInfo;
+import org.eyeseetea.malariacare.domain.entity.Credentials;
+import org.eyeseetea.malariacare.domain.entity.Device;
+import org.eyeseetea.malariacare.domain.entity.Program;
 import org.eyeseetea.malariacare.domain.exception.ApiCallException;
 import org.eyeseetea.malariacare.domain.usecase.push.PushUseCase;
-import org.eyeseetea.malariacare.rules.PreferencesDependenciesRules;
+import org.eyeseetea.malariacare.domain.usecase.push.SurveysThresholds;
+import org.eyeseetea.malariacare.presentation.executors.UIThreadExecutor;
+import org.eyeseetea.malariacare.rules.MockWebServerRule;
 import org.eyeseetea.malariacare.utils.Constants;
 import org.junit.Assert;
 import org.junit.Before;
@@ -30,6 +50,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import java.io.IOException;
+import java.util.Date;
 import java.util.List;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -45,24 +66,17 @@ public class PushUseCaseShould {
     IProgramRepository programRepository;
 
     @Rule
-    public PreferencesDependenciesRules preferencesDependenciesRules = new PreferencesDependenciesRules();
+    public MockWebServerRule mockWebServerRule = new MockWebServerRule(new AssetsFileReader());
 
     private static final String PUSH_RESPONSE_OK_ONE_SURVEY = "push_response_ok_one_survey.json";
     private static final String API_AVAILABLE_OK = "api_available_ok.json";
     private static final String API_AVAILABLE_NO_OK = "api_available_no_ok.json";
-    private Context mContext;
-
-    @Before
-    public void cleanUp() throws IOException {
-        mContext = InstrumentationRegistry.getTargetContext();
-    }
-
 
     @Test
     public void call_on_push_error_when_ape_is_not_available_during_push() throws IOException {
-        PushUseCase pushUseCase = preferencesDependenciesRules.givenPushUseCase(appInfoDataSource, deviceDataSource, mCredentialsRepository, programRepository);
-        preferencesDependenciesRules.getMockServer().enqueueMockResponseFileName(200, API_AVAILABLE_NO_OK);
-        preferencesDependenciesRules.getMockServer().enqueueMockResponseFileName(209, PUSH_RESPONSE_OK_ONE_SURVEY);
+        PushUseCase pushUseCase = givenPushUseCase();
+        mockWebServerRule.getMockServer().enqueueMockResponseFileName(200, API_AVAILABLE_NO_OK);
+        mockWebServerRule.getMockServer().enqueueMockResponseFileName(209, PUSH_RESPONSE_OK_ONE_SURVEY);
         final SurveyDB surveyDB = new SurveyDB(new OrgUnitDB(""), new ProgramDB(""),
                 new UserDB("", ""));
         surveyDB.setVoucherUid("1323544116");
@@ -143,9 +157,9 @@ public class PushUseCaseShould {
 
     @Test
     public void setUserCanAddSurveysToFalseOn209Response() throws IOException, InterruptedException {
-        PushUseCase pushUseCase = preferencesDependenciesRules.givenPushUseCase(appInfoDataSource, deviceDataSource, mCredentialsRepository, programRepository);
-        preferencesDependenciesRules.getMockServer().enqueueMockResponseFileName(200, API_AVAILABLE_OK);
-        preferencesDependenciesRules.getMockServer().enqueueMockResponseFileName(209, PUSH_RESPONSE_OK_ONE_SURVEY);
+        PushUseCase pushUseCase = givenPushUseCase();
+        mockWebServerRule.getMockServer().enqueueMockResponseFileName(200, API_AVAILABLE_OK);
+        mockWebServerRule.getMockServer().enqueueMockResponseFileName(209, PUSH_RESPONSE_OK_ONE_SURVEY);
         final SurveyDB surveyDB = new SurveyDB(new OrgUnitDB(""), new ProgramDB(""),
                 new UserDB("", ""));
         surveyDB.setVoucherUid("1323544116");
@@ -227,4 +241,40 @@ public class PushUseCaseShould {
         });
 
     }
+
+    public PushUseCase givenPushUseCase() {
+        WSPushController mWSPushController = givenWSPushController();
+
+        ISurveyRepository surveyRepository = new SurveyLocalDataSource();
+        IAsyncExecutor asyncExecutor = new IAsyncExecutor() {
+            @Override
+            public void run(Runnable runnable) {
+                runnable.run();
+            }
+        };
+        IMainExecutor mainExecutor = new UIThreadExecutor();
+        IOrganisationUnitRepository orgUnitRepository = new OrganisationUnitRepository();
+
+        SurveysThresholds surveysThresholds =
+                new SurveysThresholds(BuildConfig.LimitSurveysCount,
+                        BuildConfig.LimitSurveysTimeHours);
+        return new PushUseCase(mWSPushController, asyncExecutor, mainExecutor,
+                surveysThresholds, surveyRepository, orgUnitRepository);
+    }
+
+    public WSPushController givenWSPushController() {
+        ISurveyRepository surveyRepository = new SurveyLocalDataSource();
+        ISettingsRepository settingsRepository = new SettingsDataSource(
+                PreferencesState.getInstance().getContext());
+        Date date = new Date();
+        when(appInfoDataSource.getAppInfo()).thenReturn(new AppInfo("0", "0", "0", date, date));
+        when(deviceDataSource.getDevice()).thenReturn(new Device("phoneNumber", "imei", "version"));
+        when(mCredentialsRepository.getLastValidCredentials()).thenReturn(new Credentials(mockWebServerRule.getMockServer().getBaseEndpoint(), "test", "test"));
+        when(programRepository.getUserProgram()).thenReturn(new Program("code", "uid"));
+        eReferralsAPIClient eReferralsAPIClient = new eReferralsAPIClient(mockWebServerRule.getMockServer().getBaseEndpoint());
+        ConvertToWSVisitor mConvertToWSVisitor = new ConvertToWSVisitor(deviceDataSource, mCredentialsRepository, settingsRepository, appInfoDataSource,
+                new ProgramRepository(), new CountryVersionLocalDataSource());
+        return new WSPushController(mConvertToWSVisitor, eReferralsAPIClient, surveyRepository);
+    }
+
 }
