+++ app/src/androidTestEreferrals/java/org/eyeseetea/malariacare/data/sync/exporter/WSPushControllerShould.java
@@ -5,18 +5,31 @@ import static junit.framework.Assert.assertTrue;
 import static junit.framework.Assert.fail;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.Mockito.when;
 
+import org.eyeseetea.malariacare.AssetsFileReader;
+import org.eyeseetea.malariacare.data.database.datasources.CountryVersionLocalDataSource;
+import org.eyeseetea.malariacare.data.database.datasources.SettingsDataSource;
+import org.eyeseetea.malariacare.data.database.datasources.SurveyLocalDataSource;
 import org.eyeseetea.malariacare.data.database.model.OrgUnitDB;
 import org.eyeseetea.malariacare.data.database.model.ProgramDB;
 import org.eyeseetea.malariacare.data.database.model.SurveyDB;
 import org.eyeseetea.malariacare.data.database.model.UserDB;
+import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
+import org.eyeseetea.malariacare.data.repositories.ProgramRepository;
 import org.eyeseetea.malariacare.domain.boundary.IPushController;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IAppInfoRepository;
 import org.eyeseetea.malariacare.domain.boundary.repositories.ICredentialsRepository;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IDeviceRepository;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IProgramRepository;
+import org.eyeseetea.malariacare.domain.boundary.repositories.ISettingsRepository;
+import org.eyeseetea.malariacare.domain.boundary.repositories.ISurveyRepository;
+import org.eyeseetea.malariacare.domain.entity.AppInfo;
+import org.eyeseetea.malariacare.domain.entity.Credentials;
+import org.eyeseetea.malariacare.domain.entity.Device;
+import org.eyeseetea.malariacare.domain.entity.Program;
 import org.eyeseetea.malariacare.domain.exception.AvailableApiException;
-import org.eyeseetea.malariacare.rules.PreferencesDependenciesRules;
+import org.eyeseetea.malariacare.rules.MockWebServerRule;
 import org.eyeseetea.malariacare.utils.Constants;
 import org.hamcrest.CoreMatchers;
 import org.junit.Before;
@@ -28,13 +41,14 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @RunWith(MockitoJUnitRunner.class)
 public class WSPushControllerShould {
 
     @Rule
-    public PreferencesDependenciesRules preferencesDependenciesRules = new PreferencesDependenciesRules();
+    public MockWebServerRule mockWebServerRule = new MockWebServerRule(new AssetsFileReader());
 
     @Mock
     ICredentialsRepository mCredentialsRepository;
@@ -61,7 +75,7 @@ public class WSPushControllerShould {
     @Test
     public void push_callback_return_a_error_when_the_api_is_not_available()
             throws IOException {
-        WSPushController mWSPushController = preferencesDependenciesRules.givenWSPushController(appInfoDataSource, deviceDataSource, mCredentialsRepository, programRepository);
+        WSPushController mWSPushController = givenWSPushController();
 
         givenSomeTestSurveys();
 
@@ -99,7 +113,7 @@ public class WSPushControllerShould {
     @Test
     public void put_survey_status_conflict_if_failed_or_no_success_action_status()
             throws IOException {
-        WSPushController mWSPushController = preferencesDependenciesRules.givenWSPushController(appInfoDataSource, deviceDataSource, mCredentialsRepository, programRepository);
+        WSPushController mWSPushController = givenWSPushController();
         givenSomeTestSurveys();
 
         whenAPushResponseWithSomeConflictsIsReceived();
@@ -144,8 +158,7 @@ public class WSPushControllerShould {
     @Test
     public void update_quarantine_with_correct_status_when_do_push_with_some_quarantine_surveys()
             throws IOException {
-        WSPushController mWSPushController = preferencesDependenciesRules.givenWSPushController(appInfoDataSource, deviceDataSource,
-                mCredentialsRepository, programRepository);
+        WSPushController mWSPushController = givenWSPushController();
         givenSomeQuarantineTestSurveys();
         whenAQuarantineResponseWithSomeQuarantineSurveysIsReceived();
         assertTrue(SurveyDB.getAllQuarantineSurveys().size() == 4);
@@ -210,16 +223,16 @@ public class WSPushControllerShould {
     }
 
     private void whenAQuarantineResponseWithSomeQuarantineSurveysIsReceived() throws IOException {
-        preferencesDependenciesRules.getMockServer().enqueueMockResponse(QUARANTINE_RESPONSE_CONFLICT);
+        mockWebServerRule.getMockServer().enqueueMockResponse(QUARANTINE_RESPONSE_CONFLICT);
     }
 
     private void whenAPushResponseWithSomeConflictsIsReceived() throws IOException{
-        preferencesDependenciesRules.getMockServer().enqueueMockResponse(API_AVAILABLE_OK);
-        preferencesDependenciesRules.getMockServer().enqueueMockResponse(PUSH_RESPONSE_CONFLICT);
+        mockWebServerRule.getMockServer().enqueueMockResponse(API_AVAILABLE_OK);
+        mockWebServerRule.getMockServer().enqueueMockResponse(PUSH_RESPONSE_CONFLICT);
     }
     private void whenAPushHasNotAvailableApi() throws IOException{
-        preferencesDependenciesRules.getMockServer().enqueueMockResponse(API_AVAILABLE_NO_OK);
-        preferencesDependenciesRules.getMockServer().enqueueMockResponse(PUSH_RESPONSE_CONFLICT);
+        mockWebServerRule.getMockServer().enqueueMockResponse(API_AVAILABLE_NO_OK);
+        mockWebServerRule.getMockServer().enqueueMockResponse(PUSH_RESPONSE_CONFLICT);
     }
 
     private void deleteTestConflictGeneratedValues() {
@@ -227,5 +240,20 @@ public class WSPushControllerShould {
             SurveyDB.findById(surveyId).delete();
         }
         ProgramDB.findByUID("testProgramId").delete();
+    }
+
+    public WSPushController givenWSPushController() {
+        ISurveyRepository surveyRepository = new SurveyLocalDataSource();
+        ISettingsRepository settingsRepository = new SettingsDataSource(
+                PreferencesState.getInstance().getContext());
+        Date date = new Date();
+        when(appInfoDataSource.getAppInfo()).thenReturn(new AppInfo("0", "0", "0", date, date));
+        when(deviceDataSource.getDevice()).thenReturn(new Device("phoneNumber", "imei", "version"));
+        when(mCredentialsRepository.getLastValidCredentials()).thenReturn(new Credentials(mockWebServerRule.getMockServer().getBaseEndpoint(), "test", "test"));
+        when(programRepository.getUserProgram()).thenReturn(new Program("code", "uid"));
+        eReferralsAPIClient eReferralsAPIClient = new eReferralsAPIClient(mockWebServerRule.getMockServer().getBaseEndpoint());
+        ConvertToWSVisitor mConvertToWSVisitor = new ConvertToWSVisitor(deviceDataSource, mCredentialsRepository, settingsRepository, appInfoDataSource,
+                new ProgramRepository(), new CountryVersionLocalDataSource());
+        return new WSPushController(mConvertToWSVisitor, eReferralsAPIClient, surveyRepository);
     }
 }
